chore: Rename `VerifyShareDialog` to `ClaimPoolTweetPointsDialog`
This helps us describe the call to action as "Claim" rather than "Verify",
which more accurately describes what the user is doing.

Also, the `claimPoolTweetPoints` util was updated to handle the unwrapping
of the axios response, as well as return `res.data`. Just fewer things for
UI components to worry about.