feat: Add enhanced wallet hooks
Using these we can prompt the user to connect their wallet and execute a
callback after they have done so. We have created two variants for this.
One waits merely for the wallet to be connected, and the other waits for
the wallet to be connected and the user profile to be loaded.

The two variants allow us to use this better UX in places where there
is/isn't `UserProfileContext` available.