chore: Refactor if condition, and correct `isAuthFlowPending` variable name
The if condition in `useConnectWalletThen` now runs regardless of whether
callback is set. Makes no difference functionally, but it is clearer to
interpret. It also matches the if condition in `useConnectProfileThen`.

Also, the `isAuthFlowOver` was renamed to `isAuthFlowPending`. This more
correctly represents the meaning of the variable. The comments annotating
this variable were also updated.